refactor insight dto
