+++ types/insight/dto/deepdetect-api/index.ts
