+++ types/insight/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './deepdetect-api';
