+++ types/insight/index.ts
@@ -1,3 +1,3 @@
 export * from './schema';
 export * from './enums';
-export * from './deepdetect-api/';
+export * from './dto';
